phase2/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1

psycopg[binary]
python-dotenv==1.0.0
PyJWT==2.8.0
openai>=1.0.0
